Open TileDBGroupExp

Functional interface that initialises a TileDBGroupExp instance and opens a group object at READ or WRITE mode.

Usage

tdb_group(uri, mode = "READ", ctx = NULL, tiledb_timestamp = NULL)

Arguments

uri

URI path for the TileDB object.

mode

Mode to open : either "READ" (default) or "WRITE".

ctx

Optional tiledb::tiledb_ctx() object.

tiledb_timestamp

Set a TileDB timestamp range that the resource will be opened at. Effective only for "READ" mode. Valid options:

A NULL value (default)

An R object coercible to POSIXct with length 1 which used for end timestamp, or length 2 with start, end timestamps

An object of class tiledb_timestamp. See set_tiledb_timestamp()

Also, it can be set through active field $tiledb_timestamp.

Value

A TileDBGroupExp, R6 object.

Active bindings

ctx : A TileDB Context. See tiledb::tiledb_ctx()

tiledb_timestamp : A TileDB timestamp range that the resource will be opened at. See set_tiledb_timestamp()

uri : The URI of the TileDB object

mode: Get the mode of the object: one of the following: "CLOSED", "READ" or "WRITE"

object_type : The TileDB object type: "ARRAY","GROUP" or "INVALID"

object : Access the underlying tiledb::tiledb_group() object

members: Access the list of group members.

size : Directory size

Methods

Public Methods

$new()

$class()

$is_open()

$exists()

$get_metadata()

$set_metadata()

$create()

$open()

$close()

$remove()

$delete()

$count_members()

$get_members_df()

$get_member()

$set_member()

$names()

$member_exists()

$print()

$dump()

$has_non_members()

$non_members()

$prune_non_members()

$delete_group()

$walk_group()

$dir_tree()

Examples

if (FALSE) { # \dontrun{
 # uri path
 uri <- tempdir()

 obj <- TileDBGroup$new(uri)

 obj$create()

 obj$close()

 # new instance
 newobj <- tdb_group(uri)

 newobj$is_open() # TRUE

 newobj$mode # "READ"
} # }